Recurse over arrays for watch notices
Currently the elements of arrays inside watch notices aren't converted
to intances of `RecursiveOpenStruct`, they are converted to instances of
`Hash` instead. This mean that accessing certain fields works
differently for objects retrieved directly than for objects received in
a notice. For example, to extract the internal IP address of a node
retrieved directly the code is like this:

```ruby
ip = node.status.addresses.detect { |x| x.type == 'InternalIP' }.address
```

But for a notice it is like this:

```ruby
ip = node.status.addresses.detect { |x| x['type'] == 'InternalIP' }['address']
```

This complicates code that needs to treat objects in the same way,
regardless of how they have been retrieved.

To avoid that this patch changes the `WatchNotice` class so that the
constructor adds the RecursiveOpenStruct `:recurse_over_arrays` option.